+++ src/Site/Redirects/TargetUrlDoesNotExist.php
@@ -1,0 +1,9 @@
+<?php
+declare(strict_types=1);
+
+namespace Thinktomorrow\Chief\Site\Redirects;
+
+final class TargetUrlDoesNotExist extends \Exception
+{
+
+}