+++ CHANGELOG.md
@@ -4,6 +4,7 @@ All Notable changes to the `chief` application template will be documented in th
 principles.
 
 ## unreleased
+- Added: import redirect script. Run as `php artisan chief:import-redirects <csv>`.
 
 ## 0.7.15 - 2022-12-02
 - Fixed: honour order of selected options in window